+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/InAppWebViewFlutterPlugin.java
@@ -24,9 +24,8 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.BinaryMessenger;
-import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugin.platform.PlatformViewRegistry;
-import io.flutter.view.FlutterView;
+import io.flutter.embedding.android.FlutterView;
 
 public class InAppWebViewFlutterPlugin implements FlutterPlugin, ActivityAware {
 
@@ -64,7 +63,6 @@ public class InAppWebViewFlutterPlugin implements FlutterPlugin, ActivityAware {
   public ProcessGlobalConfigManager processGlobalConfigManager;
   public FlutterWebViewFactory flutterWebViewFactory;
   public Context applicationContext;
-  public PluginRegistry.Registrar registrar;
   public BinaryMessenger messenger;
   public FlutterPlugin.FlutterAssets flutterAssets;
   @Nullable
@@ -75,14 +73,6 @@ public class InAppWebViewFlutterPlugin implements FlutterPlugin, ActivityAware {
   public FlutterView flutterView;
 
   public InAppWebViewFlutterPlugin() {}
-
-  @SuppressWarnings("deprecation")
-  public static void registerWith(PluginRegistry.Registrar registrar) {
-    final InAppWebViewFlutterPlugin instance = new InAppWebViewFlutterPlugin();
-    instance.registrar = registrar;
-    instance.onAttachedToEngine(
-            registrar.context(), registrar.messenger(), registrar.activity(), registrar.platformViewRegistry(), registrar.view());
-  }
 
   @Override
   public void onAttachedToEngine(FlutterPluginBinding binding) {